+++ Sources/FeatherRelationalDatabase/Component/RelationalDatabaseComponentError.swift
@@ -1,0 +1,12 @@
+//
+//  SQLDatabaseComponentError.swift
+//  FeatherSQLDatabase
+//
+//  Created by Tibor Bodecs on 2023. 01. 16..
+//
+
+/// sql-database component error
+public enum RelationalDatabaseComponentError: Error {
+
+    case unknown(Error)
+}
