--- Sources/FeatherRelationalDatabase/Service/RelationalDatabaseServiceError.swift
@@ -1,12 +1,0 @@
-//
-//  SQLDatabaseServiceError.swift
-//  FeatherSQLDatabase
-//
-//  Created by Tibor Bodecs on 2023. 01. 16..
-//
-
-/// sql-database service error
-public enum RelationalDatabaseServiceError: Error {
-
-    case unknown(Error)
-}
